Move node impl struct to separate header
